+++ src/main/kotlin/se/zensum/jwt/Feature.kt
@@ -1,7 +1,5 @@
 package se.zensum.jwt
 
-import com.auth0.jwk.JwkProviderBuilder
-import com.auth0.jwt.JWTVerifier
 import com.auth0.jwt.interfaces.DecodedJWT
 import io.ktor.application.ApplicationCall
 import io.ktor.application.ApplicationCallPipeline
@@ -13,71 +11,17 @@ import io.ktor.util.AttributeKey
 
 private val REQUEST_KEY = AttributeKey<DecodedJWT>("jwt")
 
-
-class Configuration internal constructor() {
-    private var jwkURL: String? = null
-    private var issuer: String? = null
-    private var audience: String? = null
-    private var configOverride: JWTConfig? = null
-    private var providerOverride: JWTProvider? = null
-
-    private fun jwkURL() =
-        jwkURL ?: getEnv("JWK_URL")
-    private fun issuer(): String =
-        issuer ?: getEnv("JWT_ISSUER")
-    private fun audience(): String =
-        audience ?: getEnv("JWT_AUDIENCE", "")
-
-    private fun throwIfConfigSet() {
-        if (configOverride != null || providerOverride != null) {
-            throw UnsupportedOperationException("cannot set config after setting override")
-        }
-    }
-    fun jwkURL(url: String) {
-        throwIfConfigSet()
-        this.jwkURL = url
-    }
-
-    fun issuer(iss: String) {
-        throwIfConfigSet()
-        this.issuer = iss
-    }
-
-    fun audience(aud: String) {
-        throwIfConfigSet()
-        this.audience = aud
-    }
-
-    fun jwtConfig(jwtConfig: JWTConfig) {
-        if (jwkURL != null || issuer != null || audience != null) {
-            throw UnsupportedOperationException("Cannot set jwtConfig after setting anything else!")
-        }
-        this.configOverride = jwtConfig
-    }
-
-    fun jwtProvider(jwtProvider: JWTProvider) {
-        if (jwkURL != null || issuer != null || audience != null || configOverride != null) {
-            throw UnsupportedOperationException("Cannot set jwtProvider after setting anything else!")
-        }
-        this.providerOverride = jwtProvider
-    }
-
-    private fun jwkProvider() =
-        JwkProviderBuilder(jwkURL()).build()
-
-    private fun getConfig() =
-        configOverride ?: JWTConfig(jwkProvider(), issuer(), audience())
-    internal fun getProvider(): JWTProvider =
-        providerOverride ?: JWTProviderImpl(getConfig().verifier)
-}
-
 private const val BEARER_AUTH_TYPE = "Bearer "
+
+private fun extractBearer(x: String?): String? = x
+    ?.takeIf { it.startsWith(BEARER_AUTH_TYPE) }
+    ?.removePrefix(BEARER_AUTH_TYPE)
+    ?.trim()
+
 class JWTFeature internal constructor(private val provider: JWTProvider) {
     private fun intercept(context: PipelineContext<Unit, ApplicationCall>) {
         context.call.apply {
-            request.authorization()?.takeIf {
-                it.startsWith(BEARER_AUTH_TYPE)
-            }?.removePrefix(BEARER_AUTH_TYPE)?.trim()?.let {
+            extractBearer(request.authorization())?.let {
                 provider.verifyJWT(it)
             }?.let {
                 attributes.put(REQUEST_KEY, it)
